resolve circular header reference
