+++ mazerunner-core/src/adc_atmega328.cpp
@@ -4,7 +4,7 @@
  * File Created: Wednesday, 26th October 2022 12:11:36 am                     * 
  * Author: Peter Harrison                                                     * 
  * -----                                                                      * 
- * Last Modified: Thursday, 27th October 2022 11:06:47 am                     * 
+ * Last Modified: Thursday, 27th October 2022 1:32:56 pm                      * 
  * -----                                                                      * 
  * Copyright 2022 - 2022 Peter Harrison, Micromouseonline                     * 
  * -----                                                                      * 
@@ -55,75 +55,8 @@
  */
 
 #ifndef ARDUINO_ARCH_AVR
-#warning this is not an AVR
+#error This is only meant to run on an an AVR processor
 #endif
-
-void adc_isr(IAnalogueConverter &adc) {
-
-  switch (adc.m_phase) {
-    case 0: { // initialisation
-      adc.m_group_index = 0;
-      adc.m_channel_index = 0;
-      adc.m_phase = 1;
-      adc.start_conversion(15);
-    } break;
-    case 1: { // all channels get read 'dark' first
-      adc.m_adc_reading[adc.m_channel_index] = adc.get_adc_result();
-      adc.m_channel_index += 1;
-      if (adc.m_channel_index >= adc.MAX_CHANNELS) {
-        if (adc.m_emitters_enabled) {
-          adc.m_channel_index = 0;
-          adc.m_phase = 2;
-          adc.emitter_on(adc.m_emitter_pin[0]);
-          adc.start_conversion(15); // dummy conversion to get to the next isr
-        } else {
-          adc.end_conversion_cycle(); // finish the cycle
-        }
-        break;
-      }
-      adc.start_conversion(adc.m_channel_index);
-    } break;
-    case 2: { // start the first lit group
-      int channel = adc.m_group[0][adc.m_channel_index];
-      adc.start_conversion(channel);
-      adc.m_phase = 3;
-    } break;
-    case 3: { // first group conversions
-      int channel = adc.m_group[0][adc.m_channel_index];
-      adc.m_adc_reading[channel] = adc.get_adc_result() - adc.m_adc_reading[channel];
-      adc.m_channel_index += 1;
-      //            std::cout << " [" << int(adc.m_channel_index) << "] ";
-      if (adc.m_channel_index >= adc.m_group[0].size()) {
-        adc.m_channel_index = 0;
-        adc.m_phase = 4;
-        adc.emitter_off(adc.m_emitter_pin[0]);
-        adc.emitter_on(adc.m_emitter_pin[1]);
-        adc.start_conversion(15); // dummy conversion to delay one cycle
-        break;
-      }
-      channel = adc.m_group[0][adc.m_channel_index];
-      adc.start_conversion(channel);
-    } break;
-    case 4: { // start the second group
-      int channel = adc.m_group[1][adc.m_channel_index];
-      adc.start_conversion(channel);
-      adc.m_phase = 5;
-    } break;
-    case 5: { // second group conversions
-      int channel = adc.m_group[0][adc.m_channel_index];
-      adc.m_adc_reading[channel] = adc.get_adc_result() - adc.m_adc_reading[channel];
-      adc.m_channel_index += 1;
-      if (adc.m_channel_index >= adc.m_group[0].size()) {
-        adc.m_channel_index = 0;
-        adc.emitter_off(adc.m_emitter_pin[1]);
-        adc.end_conversion_cycle();
-        break;
-      }
-      channel = adc.m_group[0][adc.m_channel_index];
-      adc.start_conversion(channel);
-    } break;
-  }
-}
 
 ISR(ADC_vect) {
   adc_isr(adc);
