+++ mazerunner-core/src/adc_atmega328.h
@@ -4,7 +4,7 @@
  * File Created: Wednesday, 26th October 2022 10:51:51 pm                     *
  * Author: Peter Harrison                                                     *
  * -----                                                                      *
- * Last Modified: Thursday, 27th October 2022 11:06:32 am                     *
+ * Last Modified: Thursday, 27th October 2022 1:25:37 pm                      *
  * -----                                                                      *
  * Copyright 2022 - 2022 Peter Harrison, Micromouseonline                     *
  * -----                                                                      *
@@ -20,6 +20,7 @@
 #include "adc_atmega328.h"
 #include "digitalWriteFast.h"
 #include <Arduino.h>
+#include <wiring_private.h>
 
 class adc_atmega328 : public IAnalogueConverter {
